pyproject.toml
[project]
name = "conson-xp"
dynamic = ["version"]
description = "XP Protocol Communication Tools"
authors = [
    {name = "ldvchosal", email = "ldvchosal@github.com"},
]
dependencies = [
    "click>=8.3",
    "click-help-colors>=0.9",
    "pyyaml>=6.0",
    "structlog>=22.0",
    "fastapi>=0.118",
    "uvicorn>=0.37",
    "pydantic>=2.11",
    "HAP-python[QRCode]>=5.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {file = "LICENSE"}

[project.urls]
"Homepage" = "https://github.com/lduchosal/xp"
"Bug Reports" = "https://github.com/lduchosal/xp/issues"
"Source" = "https://github.com/lduchosal/xp"

[project.scripts]
xp = "xp.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.4",
    "pytest-cov>=7.0",
    "black>=25.9",
    "flake8>=7.3",
    "mypy>=1.18",
    "ruff>=0.13.0",
    "vulture>=2.14",
    "httpx>=0.24.0",
    "twine>=6.2.0",
    "types-PyYAML",
    "refurb>=2.2.0",
    "pdm-backend>=2.4.5",
    "isort>=6.0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=src/xp",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=60"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "reverseproxy: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src/xp"]
omit = ["tests/*", "src/xp/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true
disable_error_code = ["method-assign"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pdm]
distribution = true
version = {source = "file", path = "src/xp/__init__.py"}

[tool.pdm.scripts]
test = "env PYTHONPATH=src python -m pytest tests/ -v --cov=src/xp --cov-report=term-missing"
test-quick = "env PYTHONPATH=src python -m pytest tests/ --tb=no -q"
test-ci = "env PYTHONPATH=src python -m pytest tests/ -v --cov=src/xp --cov-branch --cov-report=xml -m 'not reverseproxy'"
test-unit = "env PYTHONPATH=src python -m pytest tests/unit/ -v"
test-integration = "env PYTHONPATH=src python -m pytest tests/integration/ -v"
test-cov = "env PYTHONPATH=src python -m pytest tests/ -v --cov=src/xp --cov-report=html --cov-report=term-missing"
lint = "ruff check --fix src/ tests/"
format = "black src/ tests/"
typecheck = "mypy src/ tests/"
vulture = "vulture src/ tests/"
refurb = "refurb src/ tests/"
isort = "isort src/ tests/"
absolufy = {shell = "find src tests -name '*.py' -type f -exec absolufy-imports --application-directories=src:tests {} +"}
check = ["lint", "format", "typecheck", "refurb", "deadvar", "test"]
bump-version = "pdm bump -v patch"
env = "echo ${PATH}"
install = "pdm install"
install-dev = "pdm install -G dev"
clean = "rm -rf .coverage coverage.xml htmlcov/ .pytest_cache/ __pycache__/ src/xp/__pycache__/"
version-show = "python -c 'import xp; print(xp.__version__)'"
version-patch = "pdm bump patch"
version-minor = "pdm bump minor"
version-major = "pdm bump major"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.vulture]
exclude = ["src/xp/cli"]
ignore_decorators = ["@click"]
make_whitelist = true
min_confidence = 80
paths = ["src", "tests"]
sort_by_size = true
verbose = false

[tool.pdm-bump]
version-files = [
    "src/xp/__init__.py:__version__",
]